adiciona cadastro de Livro

+++ src/main/java/br/com/zup/orangetalents/casadocodigo/categoria/repository/CategoriaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.com.zup.orangetalents.casadocodigo.categoria.model.Categoria;
 
-public interface CategoriaRepository extends CrudRepository<Categoria, Integer> {
+public interface CategoriaRepository extends CrudRepository<Categoria, Long> {
 
 	Optional<Categoria> findByNome(String nome);
 }
